CONFIG: Update cmake script to compile output to root
Static libraries generated will be places in /out/lib for the
SK generator to look at when creating the C dylib.

Test binaries are in /bin

In addition, the commit aligns dependencies and variablises
directories in anticipation for potential folder restructure
(i.e., removing the root coresdk and placing its content
in root).

Both /bin and /out are ignored by git