Include SDL in tests to provide SDL_main
When changing over to using pkg-config for configuring the windows
builds, it added the -mwindows link flag. This causes the compiler to
load the problem via WinMain rather then main.

However SDL2 is able to provide a WinMain which does windows specific
setup then calls main as expected on unix system. In order to use it we
need to include the SDL header where the test main functions are
defined.